requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.6
websocket-client>=1.7.0
requests>=2.31.0
yfinance>=0.2.36
feedparser>=6.0.11
gunicorn>=21.2.0
numpy>=1.26.0
eventlet>=0.35.0
